% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar_calc.R
\name{solar_calc}
\alias{solar_calc}
\title{solar_calc}
\usage{
solar_calc(latitude, longitude, excel_date, time_zone, return_var = "both")
}
\arguments{
\item{latitude}{Latitude in decimal degrees.}

\item{longitude}{Longitude in decimal degrees.}

\item{excel_date}{Date in MS Excel format. See https://github.com/rasenior/ConvertDateTime for functions to convert dates into Excel format.}

\item{time_zone}{Time zone (numeric), taking account of Daylight saving time where applicable.}

\item{return_var}{Variables to return: 'both', 'sunrise' or 'sunset'. Defaults to 'both'.}
}
\value{
Either single sunset or sunrise value expressed as minutes of the day, or a dataframe with both variables combined.
}
\description{
Calculate the time of sunrise and sunset based on date and location.
}
\examples{
solar_calc(latitude = 53.6955847, longitude = -2.0015868, excel_date = 42913, time_zone = 1, return_var = 'both')
}
